fix(app): rename call screen to active call
